+++ Frameworks/MABookmarksBar/MABookmarksBar/MABookmarksBar.h
@@ -1,0 +1,18 @@
+//
+//  MABookmarksBar.h
+//  MABookmarksBar
+//
+//  Created by Ashwin Paudel on 2022-02-15.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for MABookmarksBar.
+FOUNDATION_EXPORT double MABookmarksBarVersionNumber;
+
+//! Project version string for MABookmarksBar.
+FOUNDATION_EXPORT const unsigned char MABookmarksBarVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <MABookmarksBar/PublicHeader.h>
+
+