+++ Changes.md
@@ -1,6 +1,7 @@
 #  Changes
 
 ## v1.2
+- Created a basic bookmarks bar
 - Fixed problem where removing the webView's observer would cause an error
 - Fixed error where being on a tab and trying to close another tab, will not close the webview
 - Created a `enableBatterySaver` function in the webView